Move some common functions to rocketchat:utils
