+++ packages/rocketchat-utils/client/index.js
@@ -2,6 +2,8 @@ import { t, isRtl } from '../lib/tapi18n';
 import { isChrome, isFirefox } from './lib/browsers';
 import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
 import { Info } from '../rocketchat.info';
+import { handleError } from './lib/handleError';
+import { getUserPreference } from '../lib/getUserPreference';
 
 export {
 	t,
@@ -10,4 +12,6 @@ export {
 	isFirefox,
 	getDefaultSubscriptionPref,
 	Info,
+	handleError,
+	getUserPreference,
 };
