+++ packages/rocketchat-utils/server/index.js
@@ -1,10 +1,12 @@
 import { t, isRtl } from '../lib/tapi18n';
 import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
 import { Info } from '../rocketchat.info';
+import { getUserPreference } from '../lib/getUserPreference';
 
 export {
 	t,
 	isRtl,
 	getDefaultSubscriptionPref,
 	Info,
+	getUserPreference,
 };
